Add check to directory to prevent failing dependencies when clones have already happened.

+++ Rakefile
@@ -1,12 +1,13 @@
 ################################################################################
 # Constants
 ################################################################################
-TRELLIS_REPO    = "https://github.com/roots/trellis.git"
-TRELLIS_FOLDER  = "provision"
-BEDROCK_REPO    = "https://github.com/roots/bedrock.git"
-BEDROCK_FOLDER  = "site"
+ROOT_PATH       = File.expand_path("../", __FILE__)
+TRELLIS_FOLDER  = File.join(ROOT_PATH, "provision")
+BEDROCK_FOLDER  = File.join(ROOT_PATH, "site")
+
 BASE_REPO       = "https://github.com/robyurkowski/trellis-template.git"
-BASE_FOLDER     = "./"
+TRELLIS_REPO    = "https://github.com/roots/trellis.git"
+BEDROCK_REPO    = "https://github.com/roots/bedrock.git"
 
 
 ################################################################################
@@ -36,20 +37,24 @@ end
 namespace :bootstrap do
   desc "Cleans and preps base folder after a successful clone"
   task :base do
-    `rm -rf #{BASE_FOLDER}/.git`
+    `rm -rf #{ROOT_PATH}/.git`
     `git init`
   end
 
   desc "Downloads trellis."
   task :trellis do
-    `git clone --depth=1 #{TRELLIS_REPO} #{TRELLIS_FOLDER}`
-    `rm -rf #{TRELLIS_FOLDER}/.git`
+    unless File.exist?(TRELLIS_FOLDER)
+      `git clone --depth=1 #{TRELLIS_REPO} #{TRELLIS_FOLDER}`
+      `rm -rf #{TRELLIS_FOLDER}/.git`
+    end
   end
 
   desc "Downloads bedrock."
   task :bedrock do
-    `git clone --depth=1 #{BEDROCK_REPO} #{BEDROCK_FOLDER}`
-    `rm -rf #{BEDROCK_FOLDER}/.git`
+    unless File.exist?(BEDROCK_FOLDER)
+      `git clone --depth=1 #{BEDROCK_REPO} #{BEDROCK_FOLDER}`
+      `rm -rf #{BEDROCK_FOLDER}/.git`
+    end
   end
 
   desc "Injects non-default modules and variables into trellis files."
@@ -107,6 +112,6 @@ namespace :update do
 
   desc "Updates base rakefile."
   task :base do
-    branch_and_sync(repo: BASE_REPO, dest: BASE_FOLDER)
+    branch_and_sync(repo: BASE_REPO, dest: ROOT_PATH)
   end
 end
